server/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.0.0
passlib==1.7.4
SQLAlchemy==2.0.23
bcrypt==4.0.1
gunicorn==21.2.0 